Code refactor: home and thread list page components
