+++ forumproject/forumproject-ui/src/Routing/PrivateRoutes.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Route, Switch, Redirect } from 'react-router-dom';
 
-import Home from 'pages/Home';
-import ThreadList from 'pages/ThreadList';
+import HomePage from 'pages/HomePage';
+import ThreadListPage from 'pages/ThreadListPage';
 import CreateThread from 'pages/CreateThread';
 import Thread from 'pages/Thread';
 import Profile from 'pages/Profile';
@@ -14,9 +14,9 @@ import NotFound from 'pages/NotFound';
 const PrivateRoutes = () => {
   return (
     <Switch>
-      <Route path="/" exact component={Home} />
+      <Route path="/" exact component={HomePage} />
       <Route path="/categories" exact render={() => <Redirect to="/" />} />
-      <Route path="/categories/:categoryId" exact component={ThreadList} />
+      <Route path="/categories/:categoryId" exact component={ThreadListPage} />
       <Route
         path="/categories/:categoryId/threads/new"
         exact
