+++ forumproject/forumproject-ui/src/pages/ThreadListPage/ThreadList/style.js
@@ -4,47 +4,6 @@ import { Link } from 'react-router-dom';
 import { theme, SVGIcon } from 'layout';
 import { ReactComponent as Picture } from 'img/start_thread.svg';
 
-export const LinkButton = styled(Link)`
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 100%;
-  min-height: 3.5rem;
-  font-size: 1.6rem;
-  font-weight: 600;
-  text-align: center;
-  padding: 0 2rem;
-  color: #374350;
-  border-color: #4d5863;
-  border-style: solid;
-  border-width: 0.2rem;
-  transition: 0.3s;
-  box-shadow: 0 0 0.4rem rgba(0, 0, 0, 0.2);
-  text-decoration: none;
-  cursor: pointer;
-
-  &:hover {
-    color: #000;
-    border-color: ${theme.colors.black};
-    box-shadow: 0 0 0.6rem rgba(0, 0, 0, 0.35);
-  }
-
-  ${theme.media.minLandscapePhone} {
-    min-width: 20rem;
-    width: auto;
-    margin-left: auto;
-  }
-`;
-
-export const LinkWrapper = styled.div`
-  display: flex;
-  margin: 4rem 0;
-`;
-
-export const ThreadListWrapper = styled.div`
-  margin: 4rem 0;
-`;
-
 export const ThreadWrapper = styled.div`
   background-color: ${theme.colors.white};
   padding: 1.5rem;
@@ -117,12 +76,6 @@ export const FooterSpan = styled.span`
   font-size: 1.2rem;
   color: ${theme.colors.neutralMidDark};
   margin-right: 1rem;
-`;
-
-export const PaginationWrapper = styled.div`
-  display: flex;
-  justify-content: flex-start;
-  margin: 1rem 0 1rem;
 `;
 
 export const StartThreadPicture = styled(Picture)`
